THRIFT-5563: fix deprecation and enable xlint for java library (#2575)

+++ lib/java/src/org/apache/thrift/partial/TFieldData.java
@@ -27,7 +27,7 @@ package org.apache.thrift.partial;
  */
 public class TFieldData {
   public static int encode(byte type) {
-    return (int) (type & 0xff);
+    return type & 0xff;
   }
 
   public static int encode(byte type, short id) {
